@@ -107,18 +106,12 @@
     "sinon": "6.0.0",
     "sinon-chai": "3.2.0",
     "spectrum-colorpicker": "1.8.0",
-    "stats-webpack-plugin": "0.6.2",
     "url-loader": "1.0.1",
-    "vue": "2.5.16",
+    "vue": "^2.7.10",
     "vue-loader": "14.2.1",
     "vue-multiselect": "2.1.0",
     "vue-style-loader": "4.1.0",
-    "vue-template-compiler": "2.5.16",
-    "webpack": "3.11.0",
-    "webpack-dev-middleware": "2.0.6",
-    "webpack-dev-server": "2.11.1",
-    "webpack-merge": "4.1.3",
-    "webpack-node-externals": "1.7.2"
+    "vue-template-compiler": "^2.7.10"
   },
   "dependencies": {},
   "publishConfig": {
